+++ src/components/Button/Button.react.tsx
@@ -1,14 +1,16 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
   TouchableOpacity,
   Text,
   TextStyle,
   View,
   ViewStyle,
+  Image,
 } from 'react-native';
 import { Typography } from '@styles';
 import Next from '@assets/components/Button/Next';
 import { SvgXml } from 'react-native-svg';
+import Loading from '@assets/loading.gif';
 import Styles from './Button.styles';
 
 export interface Props {
@@ -20,16 +22,20 @@ export interface Props {
   reverse?: boolean;
   link?: boolean;
   enabled?: boolean;
-  onPress: () => void;
+  blocking?: boolean;
+  onPress: () => void | Promise<void>;
   children?: JSX.Element;
   showNextIcon?: boolean;
 }
 
 const Button: React.FC<Props> = (props: Props) => {
+  const [blocked, setBlocked] = useState(false);
+
   const {
     containerStyle,
     disabledContainerStyle,
     disabledTextStyle,
+    blocking,
     onPress,
     textStyle,
     buttonText,
@@ -49,15 +55,25 @@ const Button: React.FC<Props> = (props: Props) => {
       <TouchableOpacity
         style={[
           reverse ? Styles.buttonBackgroundReverse : Styles.buttonBackground,
-          enabled ? {} : Styles.buttonBackgroundDisabled,
+          enabled && !blocked ? {} : Styles.buttonBackgroundDisabled,
           containerStyle,
-          enabled ? {} : disabledContainerStyle,
+          enabled && !blocked ? {} : disabledContainerStyle,
           Styles.shadow,
         ]}
-        activeOpacity={enabled ? 0.7 : 1.0}
-        onPress={() => {
-          if (enabled) {
-            onPress();
+        activeOpacity={enabled && !blocked ? 0.7 : 1.0}
+        onPress={async () => {
+          if (enabled && !blocked) {
+            if (blocking) {
+              setBlocked(true);
+              try {
+                await onPress();
+              } catch (err) {
+                /* should be handled in onPress */
+              }
+              setBlocked(false);
+            } else {
+              onPress();
+            }
           }
         }}
         testID="clickable"
@@ -74,6 +90,13 @@ const Button: React.FC<Props> = (props: Props) => {
           {buttonText}
         </Text>
         {nextIcon}
+        {blocked && (
+          <Image
+            testID="loading"
+            source={Loading}
+            style={{ width: 20, height: 20 }}
+          />
+        )}
       </TouchableOpacity>
     );
   }
