+++ src/views/Home/ContactSelector.react.tsx
@@ -6,7 +6,7 @@ import {
   Platform,
   FlatList,
 } from 'react-native';
-import { Icon } from '@components';
+import { Icon, Button } from '@components';
 import { AppStackParamList } from '@navigations';
 import { StackNavigationProp } from '@react-navigation/stack';
 import { Colors, Typography } from '@styles';
@@ -22,6 +22,7 @@ import { getContacts, getUser } from '@api';
 import { dropdownError } from '@components/Dropdown/Dropdown.react';
 import { Notif, NotifActionTypes } from '@store/Notif/NotifTypes';
 import { handleNotif } from '@store/Notif/NotifiActions';
+import { sleep } from '@utils';
 import Styles from './ContactSelector.styles';
 
 type ContactSelectorScreenNavigationProp = StackNavigationProp<
@@ -142,7 +143,8 @@ class ContactSelectorScreenBase extends React.Component<Props, State> {
             try {
               await getContacts();
               await getUser();
-            } catch (e) {
+            } catch (err) {
+              console.log(err);
               dropdownError({ message: i18n.t('Error.cantRefreshContacts') });
             }
             this.setState({ refreshing: false });
