+++ src/views/Login/Login.react.tsx
@@ -214,7 +214,7 @@ class LoginScreen extends React.Component<Props, State> {
                   />
                   <Text style={Typography.FONT_REGULAR}>.</Text>
                 </View>
-                {null && (
+                {true && (
                   <Button onPress={this.devSkip} buttonText="Dev Skip" />
                 )}
                 <View style={{ width: '100%', height: 100 }} />
