+++ src/views/ReferFriends/ReferFriends.react.tsx
@@ -77,7 +77,12 @@ const ReferFriendsScreen: React.FC<Props> = (props: Props) => {
         <Button
           buttonText={i18n.t('ReferFriendsScreen.done')}
           reverse
-          onPress={() => props.navigation.navigate('SingleContact')}
+          onPress={() => {
+            props.navigation.reset({
+              index: 0,
+              routes: [{ name: 'ContactSelector' }, { name: 'SingleContact' }],
+            });
+          }}
           containerStyle={{ width: '100%' }}
         />
       </View>
