+++ typings.d.ts
@@ -2,4 +2,5 @@ declare module '*.png';
 declare module '*.jpg';
 declare module '*.json';
 declare module '*.svg';
+declare module '*.gif';
 declare module 'react-native-global-props';
